add reset password feature and comment out some logs

+++ src/services/otp.service.js
@@ -2,6 +2,7 @@ import { generateOTP } from "../utils/generateOTP.js";
 import { sendMail } from "../utils/email.js";
 import { Otp } from "../models/otp.model.js";
 import bcrypt from "bcrypt"
+import { User } from "../models/user.model.js";
 
 export const sendOtpToEmail = async (email) => {
     const otp = generateOTP();
@@ -19,12 +20,14 @@ export const sendOtpToEmail = async (email) => {
     return true;
 };
 
-export const verifyOtp = async (email, enteredOtp) => {
-    const record = await Otp.findOne({ email });
 
+export const verifyOtpAndSetPassword = async (email, enteredOtp, newPassword) => {
+    const record = await Otp.findOne({ email });
     if (!record) {
         throw new Error("OTP expired or not found");
     }
+
+    console.log("--------------------------", enteredOtp, record.otp);
 
     const isMatch = await bcrypt.compare(enteredOtp, record.otp);
 
@@ -32,6 +35,14 @@ export const verifyOtp = async (email, enteredOtp) => {
         throw new Error("Invalid OTP");
     }
 
-    await Otp.deleteMany({ email }); // Clean up on success
+    const user = await User.findOne({ email });
+    if (!user) {
+        throw new Error("User not found");
+    }
+
+    user.password = newPassword;
+    await user.save();
+
+    await Otp.deleteMany({ email }); // Clean up OTPs
     return true;
 };
